@@ -16,34 +16,34 @@
     "lint": "yarn lint-prettier && yarn lint-eslint"
   },
   "dependencies": {
-    "@unocha/hpc-api-core": "^6.0.0",
+    "@unocha/hpc-api-core": "^7.0.0",
     "apollo-server-hapi": "^3.12.0",
     "bunyan": "^1.8.15",
     "class-validator": "^0.14.0",
     "graphql": "^15.7.2",
     "knex": "0.21.1",
-    "pg": "^8.10.0",
+    "pg": "^8.11.3",
     "pm2": "^5.3.0",
     "reflect-metadata": "^0.1.13",
     "ts-node": "^10.9.1",
     "type-graphql": "^1.1.1",
     "typedi": "^0.10.0",
-    "typescript": "^5.1.6"
+    "typescript": "^5.2.2"
   },
   "devDependencies": {
     "@hapi/hapi": "^20.2.1",
     "@types/bunyan": "^1.8.8",
     "@types/hapi__hapi": "^20.0.9",
-    "@types/node": "^18.16.19",
-    "@unocha/hpc-repo-tools": "^3.0.1",
-    "eslint": "^8.44.0",
+    "@types/node": "^20.8.10",
+    "@unocha/hpc-repo-tools": "^4.0.0",
+    "eslint": "^8.52.0",
     "husky": "^8.0.3",
-    "lint-staged": "^13.2.2",
-    "prettier": "2.8.8",
+    "lint-staged": "^15.0.2",
+    "prettier": "3.0.3",
     "ts-node-dev": "^2.0.0"
   },
   "engines": {
-    "node": ">=18.16.1",
+    "node": ">=18.18.2 || >=20.9.0",
     "yarn": ">=1.22.10"
   },
   "lint-staged": {
