@@ -1,18 +1,18 @@
 import bunyan from 'bunyan';
-import { LogDataAfterBunyanProcessing } from './data';
+import { type LogDataAfterBunyanProcessing } from './data';
 
 const printLevel = (obj: LogDataAfterBunyanProcessing, useColor: boolean) => {
   if (obj.level >= bunyan.FATAL) {
-    return (useColor ? '\x1b[0;31;1m' : '') + '[FATAL]';
+    return `${useColor ? '\u001B[0;31;1m' : ''}[FATAL]`;
   }
   if (obj.level >= bunyan.ERROR) {
-    return (useColor ? '\x1b[0;31;1m' : '') + '[ERROR]';
+    return `${useColor ? '\u001B[0;31;1m' : ''}[ERROR]`;
   }
   if (obj.level >= bunyan.WARN) {
-    return (useColor ? '\x1b[0;33;1m' : '') + '[WARN]';
+    return `${useColor ? '\u001B[0;33;1m' : ''}[WARN]`;
   }
   if (obj.level >= bunyan.INFO) {
-    return (useColor ? '\x1b[0;36;1m' : '') + '[INFO]';
+    return `${useColor ? '\u001B[0;36;1m' : ''}[INFO]`;
   }
   if (obj.level >= bunyan.DEBUG) {
     return '[DEBUG]';
@@ -28,23 +28,16 @@ export const printLog = (
 ) => {
   process.stdout.write(
     // Make Dim
-    (useColor ? '\x1b[2m' : '') +
+    `${
+      useColor ? '\u001B[2m' : ''
       // Time
-      '[' +
-      obj.time.getHours() +
-      ':' +
-      obj.time.getMinutes() +
-      ':' +
-      obj.time.getSeconds() +
-      ':' +
-      obj.time.getMilliseconds() +
-      '] ' +
-      printLevel(obj, useColor) +
-      ' ' +
+    }[${obj.time.getHours()}:${obj.time.getMinutes()}:${obj.time.getSeconds()}:${obj.time.getMilliseconds()}] ${printLevel(
+      obj,
+      useColor
+    )} ${
       // Reset colors
-      (useColor ? '\x1b[0m' : '') +
-      obj.msg +
-      '\n'
+      useColor ? '\u001B[0m' : ''
+    }${obj.msg}\n`
   );
   if (obj.stackTrace) {
     process.stdout.write(
