@@ -1,4 +1,4 @@
-import { Database } from '@unocha/hpc-api-core/src/db/type';
+import { type Database } from '@unocha/hpc-api-core/src/db/type';
 
 export default interface Context {
   models: Database;
