@@ -1,5 +1,5 @@
-import { PlanId } from '@unocha/hpc-api-core/src/db/models/plan';
-import { Database } from '@unocha/hpc-api-core/src/db/type';
+import { type PlanId } from '@unocha/hpc-api-core/src/db/models/plan';
+import { type Database } from '@unocha/hpc-api-core/src/db/type';
 import { NotFoundError } from '@unocha/hpc-api-core/src/util/error';
 import { createBrandedValue } from '@unocha/hpc-api-core/src/util/types';
 import { Service } from 'typedi';
